requirements.txt
fastapi==0.111.0
sqlalchemy==2.0.30
python-dotenv